C64 updates
* Update to VICE 2.4.28
* Add Virtual Device Traps option
* Add more options to System & Media Control menu including those for disk drives
* Set AutostartPrgDisk.d64 as PRG autostart filename
* Set TDE default to off
* Various menu item tweaks

+++ C64.emu/src/vice/scpu64/scpu64memsnapshot.c
@@ -174,17 +174,17 @@ int scpu64_snapshot_write_module(snapshot_t *s, int save_roms)
         return -1;
     }
 
-    if (SMW_B(m, mem_pport) < 0
-        || SMW_B(m, mem_reg_soft_1mhz) < 0
-        || SMW_B(m, mem_reg_sys_1mhz) < 0
-        || SMW_B(m, mem_reg_hwenable) < 0
-        || SMW_B(m, mem_reg_dosext) < 0
-        || SMW_B(m, mem_reg_ramlink) < 0
-        || SMW_B(m, mem_reg_optim) < 0
-        || SMW_B(m, mem_reg_bootmap) < 0
-        || SMW_B(m, mem_reg_simm) < 0
-        || SMW_B(m, export.exrom) < 0
-        || SMW_B(m, export.game) < 0
+    if (SMW_B(m, (BYTE)mem_pport) < 0
+        || SMW_B(m, (BYTE)mem_reg_soft_1mhz) < 0
+        || SMW_B(m, (BYTE)mem_reg_sys_1mhz) < 0
+        || SMW_B(m, (BYTE)mem_reg_hwenable) < 0
+        || SMW_B(m, (BYTE)mem_reg_dosext) < 0
+        || SMW_B(m, (BYTE)mem_reg_ramlink) < 0
+        || SMW_B(m, (BYTE)mem_reg_optim) < 0
+        || SMW_B(m, (BYTE)mem_reg_bootmap) < 0
+        || SMW_B(m, (BYTE)mem_reg_simm) < 0
+        || SMW_B(m, (BYTE)export.exrom) < 0
+        || SMW_B(m, (BYTE)export.game) < 0
         || scpu64_snapshot_write_cpu_state(m)
         || SMW_DW(m, mem_simm_ram_mask) < 0
         || SMW_BA(m, mem_ram, SCPU64_RAM_SIZE) < 0
